% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr6_1580213_1582559.R
\docType{data}
\name{chr6_1580213_1582559}
\alias{chr6_1580213_1582559}
\title{Sequence of human chromosome 6 from 1,580,213 until 1,582,559}
\format{The data set is a DNAStringSet object containing one sequence of length 2346 nucleotides.}
\usage{
data (chr6_1580213_1582559)
}
\description{
A DNAStringSet object containing the data of human chromosome 6 starting with position 1,580,213 and ending at 1,582,559.
}
\seealso{
\code{\link{motif_detection}}
}
\keyword{datasets}
