% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr6_1581473_1586032.R
\docType{data}
\name{chr6_1581473_1586032}
\alias{chr6_1581473_1586032}
\title{Sequence of human chromosome 6 from 1,581,473 until 1,586,032}
\format{The data set is a DNAStringSet object containing one sequence of length 4559 nucleotides.}
\usage{
data (chr6_1581473_1586032)
}
\description{
A DNAStringSet object containing the data of human chromosome 6 starting with position 1,581,473 and ending at 1,586,032.
}
\seealso{
\code{\link{motif_detection}}
}
\keyword{datasets}
